+++ packages/contracts/contracts/TokenRegistry/TokenRegistry.sol
@@ -18,7 +18,7 @@
 
 pragma solidity ^0.4.11;
 
-import "./utils/Ownable.sol";
+import "../../contracts_old/Ownable/v1.sol";
 
 /// @title Token Registry - Stores metadata associated with ERC20 tokens. See ERC22 https://github.com/ethereum/EIPs/issues/22
 /// @author Amir Bandeali - <amir@0xProject.com>, Will Warren - <will@0xProject.com>
@@ -306,4 +306,3 @@ contract TokenRegistry is Ownable {
         return tokenAddresses;
     }
 }
-
