+++ packages/contracts/contracts/UnlimitedAllowanceToken/UnlimitedAllowanceToken.sol
@@ -16,15 +16,15 @@
 
 */
 
-pragma solidity ^0.4.11;
+pragma solidity 0.4.18;
 
-import "./StandardToken.sol";
+import "../ERC20Token/ERC20Token.sol";
 
-contract UnlimitedAllowanceToken is StandardToken {
+contract UnlimitedAllowanceToken is ERC20Token {
 
     uint constant MAX_UINT = 2**256 - 1;
 
-    /// @dev ERC20 transferFrom, modified such that an allowance of MAX_UINT represents an unlimited allowance.
+    /// @dev ERC20 transferFrom, modified such that an allowance of MAX_UINT represents an unlimited allowance. See https://github.com/ethereum/EIPs/issues/717
     /// @param _from Address to transfer from.
     /// @param _to Address to transfer to.
     /// @param _value Amount to transfer.
@@ -34,20 +34,13 @@ contract UnlimitedAllowanceToken is StandardToken {
         returns (bool)
     {
         uint allowance = allowed[_from][msg.sender];
-        if (balances[_from] >= _value
-            && allowance >= _value
-            && balances[_to] + _value >= balances[_to]
-        ) {
-            balances[_to] += _value;
-            balances[_from] -= _value;
-            if (allowance < MAX_UINT) {
-                allowed[_from][msg.sender] -= _value;
-            }
-            Transfer(_from, _to, _value);
-            return true;
-        } else {
-            return false;
+        require(balances[_from] >= _value && allowance >= _value && balances[_to] + _value >= balances[_to]);
+        balances[_to] += _value;
+        balances[_from] -= _value;
+        if (allowance < MAX_UINT) {
+            allowed[_from][msg.sender] -= _value;
         }
+        Transfer(_from, _to, _value);
+        return true;
     }
 }
-
