+++ packages/contracts/contracts_old/Ownable/v1.sol
@@ -25,4 +25,3 @@ contract Ownable {
         }
     }
 }
-
