+++ packages/contracts/contracts_old/UnlimitedAllowanceToken/v1.sol
@@ -16,32 +16,37 @@
 
 */
 
-pragma solidity 0.4.18;
+pragma solidity ^0.4.11;
 
-import "./ERC20Token.sol";
+import "../../packages/StandardToken/StandardToken.sol";
 
-contract UnlimitedAllowanceToken_v2 is ERC20Token {
+contract UnlimitedAllowanceToken is StandardToken {
 
     uint constant MAX_UINT = 2**256 - 1;
 
-    /// @dev ERC20 transferFrom, modified such that an allowance of MAX_UINT represents an unlimited allowance. See https://github.com/ethereum/EIPs/issues/717
+    /// @dev ERC20 transferFrom, modified such that an allowance of MAX_UINT represents an unlimited allowance.
     /// @param _from Address to transfer from.
     /// @param _to Address to transfer to.
     /// @param _value Amount to transfer.
     /// @return Success of transfer.
     function transferFrom(address _from, address _to, uint _value)
-        public 
-        returns (bool) 
+        public
+        returns (bool)
     {
         uint allowance = allowed[_from][msg.sender];
-        require(balances[_from] >= _value && allowance >= _value && balances[_to] + _value >= balances[_to]); 
-        balances[_to] += _value;
-        balances[_from] -= _value;
-        if (allowance < MAX_UINT) {
-            allowed[_from][msg.sender] -= _value;
+        if (balances[_from] >= _value
+            && allowance >= _value
+            && balances[_to] + _value >= balances[_to]
+        ) {
+            balances[_to] += _value;
+            balances[_from] -= _value;
+            if (allowance < MAX_UINT) {
+                allowed[_from][msg.sender] -= _value;
+            }
+            Transfer(_from, _to, _value);
+            return true;
+        } else {
+            return false;
         }
-        Transfer(_from, _to, _value);
-        return true;
     }
 }
-
